+++ client/analytics/report/stock/table.js
@@ -10,8 +10,8 @@ import { Component } from '@wordpress/element';
  */
 import { Link } from '@woocommerce/components';
 import { getNewPath, getPersistedQuery } from '@woocommerce/navigation';
-import { numberFormat } from '@woocommerce/number';
-import { getSetting } from '@woocommerce/wc-admin-settings';
+import { formatValue } from 'lib/number-format';
+import { getAdminLink, getSetting } from '@woocommerce/wc-admin-settings';
 
 /**
  * Internal dependencies
@@ -85,12 +85,13 @@ export default class StockReportTable extends Component {
 				</Link>
 			);
 
+			const editProductLink = getAdminLink( 'post.php?action=edit&post=' + ( parent_id || id ) );
 			const stockStatusLink = isLowStock( stock_status, stock_quantity, low_stock_amount ) ? (
-				<Link href={ 'post.php?action=edit&post=' + ( parent_id || id ) } type="wp-admin">
+				<Link href={ editProductLink } type="wp-admin">
 					{ _x( 'Low', 'Indication of a low quantity', 'woocommerce-admin' ) }
 				</Link>
 			) : (
-				<Link href={ 'post.php?action=edit&post=' + ( parent_id || id ) } type="wp-admin">
+				<Link href={ editProductLink } type="wp-admin">
 					{ stockStatuses[ stock_status ] }
 				</Link>
 			);
@@ -109,7 +110,9 @@ export default class StockReportTable extends Component {
 					value: stock_status,
 				},
 				{
-					display: manage_stock ? numberFormat( stock_quantity ) : __( 'N/A', 'woocommerce-admin' ),
+					display: manage_stock
+						? formatValue( 'number', stock_quantity )
+						: __( 'N/A', 'woocommerce-admin' ),
 					value: stock_quantity,
 				},
 			];
@@ -121,23 +124,23 @@ export default class StockReportTable extends Component {
 		return [
 			{
 				label: _n( 'product', 'products', products, 'woocommerce-admin' ),
-				value: numberFormat( products ),
+				value: formatValue( 'number', products ),
 			},
 			{
 				label: __( 'out of stock', outofstock, 'woocommerce-admin' ),
-				value: numberFormat( outofstock ),
+				value: formatValue( 'number', outofstock ),
 			},
 			{
 				label: __( 'low stock', lowstock, 'woocommerce-admin' ),
-				value: numberFormat( lowstock ),
+				value: formatValue( 'number', lowstock ),
 			},
 			{
 				label: __( 'on backorder', onbackorder, 'woocommerce-admin' ),
-				value: numberFormat( onbackorder ),
+				value: formatValue( 'number', onbackorder ),
 			},
 			{
 				label: __( 'in stock', instock, 'woocommerce-admin' ),
-				value: numberFormat( instock ),
+				value: formatValue( 'number', instock ),
 			},
 		];
 	}