+++ src/API/Reports/Customers/DataStore.php
@@ -59,7 +59,7 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 	 */
 	protected function assign_report_columns() {
 		global $wpdb;
-		$table_name = self::get_db_table_name();
+		$table_name           = self::get_db_table_name();
 		$this->report_columns = array(
 			'id'               => "{$table_name}.customer_id as id",
 			'user_id'          => 'user_id',
@@ -74,8 +74,8 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 			'date_last_active' => 'IF( date_last_active <= "0000-00-00 00:00:00", NULL, date_last_active ) AS date_last_active',
 			'date_last_order'  => "MAX( {$wpdb->prefix}wc_order_stats.date_created ) as date_last_order",
 			'orders_count'     => 'SUM( CASE WHEN parent_id = 0 THEN 1 ELSE 0 END ) as orders_count',
-			'total_spend'      => 'SUM( gross_total ) as total_spend',
-			'avg_order_value'  => '( SUM( gross_total ) / COUNT( order_id ) ) as avg_order_value',
+			'total_spend'      => 'SUM( total_sales ) as total_spend',
+			'avg_order_value'  => '( SUM( total_sales ) / COUNT( order_id ) ) as avg_order_value',
 		);
 	}
 
@@ -121,7 +121,7 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 	 * @param array  $query_args Parameters supplied by the user.
 	 * @param string $table_name Name of the db table relevant for the date constraint.
 	 */
-	protected function get_time_period_sql_params( $query_args, $table_name ) {
+	protected function add_time_period_sql_params( $query_args, $table_name ) {
 		global $wpdb;
 
 		$this->clear_sql_clause( array( 'where', 'where_time', 'having' ) );
@@ -188,14 +188,14 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 	 *
 	 * @param array $query_args Query arguments supplied by the user.
 	 */
-	protected function get_sql_query_params( $query_args ) {
+	protected function add_sql_query_params( $query_args ) {
 		global $wpdb;
 		$customer_lookup_table  = self::get_db_table_name();
 		$order_stats_table_name = $wpdb->prefix . 'wc_order_stats';
 
-		$this->get_time_period_sql_params( $query_args, $customer_lookup_table );
+		$this->add_time_period_sql_params( $query_args, $customer_lookup_table );
 		$this->get_limit_sql_params( $query_args );
-		$this->get_order_by_sql_params( $query_args );
+		$this->add_order_by_sql_params( $query_args );
 		$this->subquery->add_sql_clause( 'left_join', "LEFT JOIN {$order_stats_table_name} ON {$customer_lookup_table}.customer_id = {$order_stats_table_name}.customer_id" );
 
 		$match_operator = $this->get_match_operator( $query_args );
@@ -259,11 +259,11 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 				'format' => '%d',
 			),
 			'total_spend'     => array(
-				'column' => 'SUM( gross_total )',
+				'column' => 'SUM( total_sales )',
 				'format' => '%f',
 			),
 			'avg_order_value' => array(
-				'column' => '( SUM( gross_total ) / COUNT( order_id ) )',
+				'column' => '( SUM( total_sales ) / COUNT( order_id ) )',
 				'format' => '%f',
 			),
 		);
@@ -352,7 +352,7 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 			);
 
 			$selections       = $this->selected_columns( $query_args );
-			$sql_query_params = $this->get_sql_query_params( $query_args );
+			$sql_query_params = $this->add_sql_query_params( $query_args );
 
 			$db_records_count = (int) $wpdb->get_var(
 				"SELECT COUNT(*) FROM (
@@ -464,7 +464,7 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 			'state'            => $order->get_billing_state( 'edit' ),
 			'postcode'         => $order->get_billing_postcode( 'edit' ),
 			'country'          => $order->get_billing_country( 'edit' ),
-			'date_last_active' => date( 'Y-m-d H:i:s', $order->get_date_created( 'edit' )->getTimestamp() ),
+			'date_last_active' => gmdate( 'Y-m-d H:i:s', $order->get_date_created( 'edit' )->getTimestamp() ),
 		);
 		$format = array(
 			'%s',
@@ -603,7 +603,7 @@ class DataStore extends ReportsDataStore implements DataStoreInterface {
 			'postcode'         => $customer->get_billing_postcode( 'edit' ),
 			'country'          => $customer->get_billing_country( 'edit' ),
 			'date_registered'  => $customer->get_date_created( 'edit' )->date( TimeInterval::$sql_datetime_format ),
-			'date_last_active' => $last_active ? date( 'Y-m-d H:i:s', $last_active ) : null,
+			'date_last_active' => $last_active ? gmdate( 'Y-m-d H:i:s', $last_active ) : null,
 		);
 		$format      = array(
 			'%d',