+++ packages/components/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Removed WC-Admin specific actions from `<TableCard />` component.
 - Export the `<CompareButton />` component.
 - Add `<TextControl />` component.
+- Require `currency` prop in `<AdvancedFilters />` component.
+- Remove call to `getAdminLink()` inside the `<Link />` component.
 
 # 4.0.0
 - Added a new `<ScrollTo />` component.